Split out definition of require_gem to standalone file
