+++ bin/crew
@@ -1,26 +1,9 @@
 #!/usr/bin/env ruby
+# We require color here to get access to .lightred.
 require_relative '../lib/color'
-# Disallow sudo
+# Disallow sudo.
 abort 'Chromebrew should not be run as root.'.lightred if Process.uid.zero?
-def require_gem(gem_name_and_require = nil, require_override = nil)
-  # Allow only loading gems when needed.
-  return if gem_name_and_require.nil?
-
-  gem_name = gem_name_and_require.split('/')[0]
-  begin
-    gem gem_name
-  rescue LoadError
-    puts " -> install #{gem_name} gem".orange
-    Gem.install(gem_name)
-    gem gem_name
-  end
-  requires = if require_override.nil?
-               gem_name_and_require.split('/')[1].nil? ? gem_name_and_require.split('/')[0] : gem_name_and_require
-             else
-               require_override
-             end
-  require requires
-end
+require_relative '../lib/require_gem'
 require_gem('activesupport', 'active_support/core_ext/object/blank')
 require_gem('highline')
 # The json gem can break when upgrading from a much older version of ruby.
