+++ lib/misc_functions.rb
@@ -1,27 +1,7 @@
 # lib/misc_functions.rb
 # Generic implementations of various functions/algorithms that are not crew-specific.
+require 'matrix'
 require_relative 'color'
-
-def require_gem(gem_name_and_require = nil, require_override = nil)
-  # Allow only loading gems when needed.
-  return if gem_name_and_require.nil?
-
-  gem_name = gem_name_and_require.split('/')[0]
-  begin
-    gem gem_name
-  rescue LoadError
-    puts " -> install #{gem_name} gem".orange
-    Gem.install(gem_name)
-    gem gem_name
-  end
-  requires = if require_override.nil?
-               gem_name_and_require.split('/')[1].nil? ? gem_name_and_require.split('/')[0] : gem_name_and_require
-             else
-               require_override
-             end
-  require requires
-end
-require_gem('matrix')
 
 class MiscFunctions
   def self.human_size(bytes)
