+++ lib/package.rb
@@ -3,27 +3,9 @@ require 'json'
 require_relative 'const'
 require_relative 'color'
 require_relative 'package_helpers'
+require_relative 'require_gem'
 require_relative 'selector'
 
-def require_gem(gem_name_and_require = nil, require_override = nil)
-  # Allow only loading gems when needed.
-  return if gem_name_and_require.nil?
-
-  gem_name = gem_name_and_require.split('/')[0]
-  begin
-    gem gem_name
-  rescue LoadError
-    puts " -> install #{gem_name} gem".orange
-    Gem.install(gem_name)
-    gem gem_name
-  end
-  requires = if require_override.nil?
-               gem_name_and_require.split('/')[1].nil? ? gem_name_and_require.split('/')[0] : gem_name_and_require
-             else
-               require_override
-             end
-  require requires
-end
 require_gem 'highline'
 require_gem 'timeout'
