+++ tools/create_default_gems_packages.rb
@@ -14,25 +14,8 @@ require 'fileutils'
 require 'json'
 require_relative '../lib/color'
 require_relative '../lib/const'
-def require_gem(gem_name_and_require = nil, require_override = nil)
-  # Allow only loading gems when needed.
-  return if gem_name_and_require.nil?
+require_relative '../lib/require_gem'
 
-  gem_name = gem_name_and_require.split('/')[0]
-  begin
-    gem gem_name
-  rescue LoadError
-    puts " -> install #{gem_name} gem".orange
-    Gem.install(gem_name)
-    gem gem_name
-  end
-  requires = if require_override.nil?
-               gem_name_and_require.split('/')[1].nil? ? gem_name_and_require.split('/')[0] : gem_name_and_require
-             else
-               require_override
-             end
-  require requires
-end
 require_gem('httpparty')
 
 def create_default_gems_package
