+++ tools/create_gem_packages.rb
@@ -13,26 +13,8 @@ require 'fileutils'
 require 'json'
 require_relative '../lib/color'
 require_relative '../lib/const'
+require_relative '../lib/require_gem'
 
-def require_gem(gem_name_and_require = nil, require_override = nil)
-  # Allow only loading gems when needed.
-  return if gem_name_and_require.nil?
-
-  gem_name = gem_name_and_require.split('/')[0]
-  begin
-    gem gem_name
-  rescue LoadError
-    puts " -> install #{gem_name} gem".orange
-    Gem.install(gem_name)
-    gem gem_name
-  end
-  requires = if require_override.nil?
-               gem_name_and_require.split('/')[1].nil? ? gem_name_and_require.split('/')[0] : gem_name_and_require
-             else
-               require_override
-             end
-  require requires
-end
 require_gem('httpparty')
 
 def check_gem_binary_build_needed(gem_name = nil, gem_version = nil)
