+++ tools/json.rb
@@ -6,26 +6,8 @@ $LOAD_PATH.unshift '../lib'
 require_relative '../lib/const'
 require_relative '../lib/package'
 require_relative '../lib/package_utils'
+require_relative '../lib/require_gem'
 
-def require_gem(gem_name_and_require = nil, require_override = nil)
-  # Allow only loading gems when needed.
-  return if gem_name_and_require.nil?
-
-  gem_name = gem_name_and_require.split('/')[0]
-  begin
-    gem gem_name
-  rescue LoadError
-    puts " -> install #{gem_name} gem".orange
-    Gem.install(gem_name)
-    gem gem_name
-  end
-  requires = if require_override.nil?
-               gem_name_and_require.split('/')[1].nil? ? gem_name_and_require.split('/')[0] : gem_name_and_require
-             else
-               require_override
-             end
-  require requires
-end
 require_gem('activesupport', 'active_support/core_ext/object/blank')
 require_gem('ptools')
